Changed saving to append, changed listener data properties to numpy arrays

+++ pythonapp/main.py
@@ -34,11 +34,15 @@ match programMode:
 
     ########## Data Collection ##########
     case 0:
-        keyboard.wait(killKey)
-        while keyboard.is_pressed(killKey):
-            time.sleep(0.1)
-        inputListener.buttonData = inputListener.buttonData[:-2] # Strip kill key from data
-        inputListener.save_to_files(dataDirectory, dataLabel)
+        while True:
+            time.sleep(10)
+            if keyboard.is_pressed(killKey):
+                while keyboard.is_pressed(killKey):
+                    time.sleep(0.5)
+                inputListener.buttonData = inputListener.buttonData[:-2] # Strip kill key from data
+                inputListener.save_to_files(dataDirectory, dataLabel)
+                break
+            inputListener.save_to_files(dataDirectory, dataLabel)
 
     ########## Model Training ##########
     case 1:
